requirements.txt
flask
wtforms
flask-wtf
werkzeug
cassandra-driver
numpy
tensorflow==1.14
